+++ src/bunit.web/TestDoubles/NavigationInterception/FakeNavigationInterception.cs
@@ -1,0 +1,11 @@
+using System;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Components.Routing;
+
+namespace Bunit.TestDoubles
+{
+	internal sealed class FakeNavigationInterception : INavigationInterception
+	{
+		public Task EnableNavigationInterceptionAsync() => Task.CompletedTask;
+	}
+}